refactor(task): use public API for IssueFlowService imports
Replace deep imports from services.issue.flow with public API access
via services.issue.__init__.py to complete Phase 6b decoupling.

Changes:
- src/vibe3/services/task/resume.py: import from vibe3.services.issue
- src/vibe3/services/task/show.py: import from vibe3.services.issue

IssueFlowService is already exported in issue/__init__.py public API,
so this is a path-only fix with no behavior changes.

Refs: #2576
