+++ src/vibe3/services/task/resume.py
@@ -14,7 +14,7 @@ from vibe3.clients import BackendProtocol, GitClient, GitHubClient
 from vibe3.exceptions import UserError
 from vibe3.models import IssueState
 from vibe3.services.flow.service import FlowService
-from vibe3.services.issue.flow import IssueFlowService
+from vibe3.services.issue import IssueFlowService
 from vibe3.services.shared.label_service import LabelService
 from vibe3.services.shared.status_query import StatusQueryService
